Improved task generation
data_seed's for tasks are now distinct also across dimensions which increases the accuracy since more train data is taken into account. We also corrected the Python script for ternary dataset generation (3gen_mnist.py). Small tweaks for sufficient accuracies per dimension in regularized_evolution.cc. The launch script 3MD.sh is modified for the changed task_util.cc: dimensions 32 and 128 add one epoch and 64 two epochs to the given num_train_epochs; the true number of tasks allocated per dimension is a multiple of the given num_tasks, as follows: 2x for 16 and 256, 3x for 32 and 128 and 4x for 64; basically Dimension 64 is the workhorse, 16 and 32 are bootstraping, 128 is the finishing and 256 the verifying dimension(s)